User authentication and some db changes

+++ backend/src/db/migrations/20240217145249_createEntriesTable.js
@@ -5,7 +5,7 @@
 exports.up = function (knex) {
     return knex.schema.createTable('entries', function (table) {
         table.increments('entry_id').primary()
-        table.integer('person_id').unsigned
+        table.integer('person_id').unsigned()
         table
             .foreign('person_id')
             .references('person_id')
@@ -28,7 +28,14 @@ exports.up = function (knex) {
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function (knex) {
-    return knex.schema.dropTable('entries')
-}
 
+exports.down = async function (knex) {
+    // Drop the foreign key constraints in 'entries' table first
+    const hasEntriesTable = await knex.schema.hasTable('entries')
+    if (hasEntriesTable) {
+        await knex.schema.table('entries', function (table) {
+            table.dropForeign('person_id')
+        })
+        await knex.schema.dropTable('entries')
+    }
+}
